Refactor user statistics handling and update imports
- Moved user-specific count functions to the `userStatsCounts.ts` file, enhancing code organization and maintainability.
- Updated imports in `aggregateQueries.ts`, `aggregateRepairs.ts`, and `aggregateWorkflows.ts` to reflect the new structure, improving clarity and reducing redundancy.
- Added `userStatsCounts` to the API definitions, ensuring all relevant modules are accessible for user statistics management.
