+++ convex/aggregateQueries.ts
@@ -27,6 +27,20 @@ import {
   totalQuestionCount,
 } from './aggregates';
 import { getCurrentUserOrThrow } from './users';
+import {
+  getUserAnsweredCount,
+  getUserAnsweredCountByGroup,
+  getUserAnsweredCountBySubtheme,
+  getUserAnsweredCountByTheme,
+  getUserBookmarksCount,
+  getUserBookmarksCountByGroup,
+  getUserBookmarksCountBySubtheme,
+  getUserBookmarksCountByTheme,
+  getUserIncorrectCount,
+  getUserIncorrectCountByGroup,
+  getUserIncorrectCountBySubtheme,
+  getUserIncorrectCountByTheme,
+} from './userStatsCounts';
 import { getWeekString } from './utils';
 
 // ----------------------------------------------------------------------------
@@ -256,198 +270,7 @@ export async function getGroupQuestionCount(
   });
 }
 
-// ============================================================================
-// USER-SPECIFIC COUNT FUNCTIONS - UPDATED TO USE userStatsCounts TABLE
-// ============================================================================
-
-/**
- * Get user answered count using userStatsCounts table (replaces getUserAnsweredCount)
- */
-export async function getUserAnsweredCount(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.totalAnswered || 0;
-}
-
-/**
- * Get user incorrect count using userStatsCounts table (replaces getUserIncorrectCount)
- */
-export async function getUserIncorrectCount(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.totalIncorrect || 0;
-}
-
-/**
- * Get user bookmarks count using userStatsCounts table (replaces getUserBookmarksCount)
- */
-export async function getUserBookmarksCount(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.totalBookmarked || 0;
-}
-
-/**
- * Get user answered count by theme using userStatsCounts table
- */
-export async function getUserAnsweredCountByTheme(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-  themeId: Id<'themes'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.answeredByTheme[themeId] || 0;
-}
-
-/**
- * Get user incorrect count by theme using userStatsCounts table
- */
-export async function getUserIncorrectCountByTheme(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-  themeId: Id<'themes'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.incorrectByTheme[themeId] || 0;
-}
-
-/**
- * Get user bookmarks count by theme using userStatsCounts table
- */
-export async function getUserBookmarksCountByTheme(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-  themeId: Id<'themes'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.bookmarkedByTheme[themeId] || 0;
-}
-
-/**
- * Get user answered count by subtheme using userStatsCounts table
- */
-export async function getUserAnsweredCountBySubtheme(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-  subthemeId: Id<'subthemes'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.answeredBySubtheme[subthemeId] || 0;
-}
-
-/**
- * Get user incorrect count by subtheme using userStatsCounts table
- */
-export async function getUserIncorrectCountBySubtheme(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-  subthemeId: Id<'subthemes'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.incorrectBySubtheme[subthemeId] || 0;
-}
-
-/**
- * Get user bookmarks count by subtheme using userStatsCounts table
- */
-export async function getUserBookmarksCountBySubtheme(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-  subthemeId: Id<'subthemes'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.bookmarkedBySubtheme[subthemeId] || 0;
-}
-
-/**
- * Get user answered count by group using userStatsCounts table
- */
-export async function getUserAnsweredCountByGroup(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-  groupId: Id<'groups'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.answeredByGroup[groupId] || 0;
-}
-
-/**
- * Get user incorrect count by group using userStatsCounts table
- */
-export async function getUserIncorrectCountByGroup(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-  groupId: Id<'groups'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.incorrectByGroup[groupId] || 0;
-}
-
-/**
- * Get user bookmarks count by group using userStatsCounts table
- */
-export async function getUserBookmarksCountByGroup(
-  ctx: QueryCtx,
-  userId: Id<'users'>,
-  groupId: Id<'groups'>,
-): Promise<number> {
-  const userCounts = await ctx.db
-    .query('userStatsCounts')
-    .withIndex('by_user', q => q.eq('userId', userId))
-    .first();
-
-  return userCounts?.bookmarkedByGroup[groupId] || 0;
-}
+// User-specific count helpers moved to `userStatsCounts.ts`.
 
 /**
  * Count questions based on filter type only (no taxonomy selection yet)
